+++ src/components/Header/BurgerMenu/styles.js
@@ -1,0 +1,8 @@
+// Material UI
+import { makeStyles } from '@material-ui/core/styles';
+
+export const useStyles = makeStyles((theme) => ({
+  menuButton: {
+    marginRight: theme.spacing(2),
+  },
+}));
