+++ src/components/Header/DesktopMenu/styles.js
@@ -1,0 +1,11 @@
+// Material UI
+import { makeStyles } from '@material-ui/core/styles';
+
+export const useStyles = makeStyles((theme) => ({
+  sectionDesktop: {
+    display: 'none',
+    [theme.breakpoints.up('md')]: {
+      display: 'flex',
+    },
+  },
+}));
