+++ src/components/Videos/__tests__/Videos.test.js
@@ -2,16 +2,16 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 
 // Components
-import Videos from './Videos';
+import VideoList from '../VideoList';
 
 // Utils
-import mockedVideos from '../../utils/youtube-videos-mock.json';
+import mockedVideos from '../../../utils/youtube-videos-mock.json';
 
 const { items } = mockedVideos;
 
 describe('Video component', () => {
   it('displays the list of the videos', () => {
-    render(<Videos />);
+    render(<VideoList />);
 
     const cardElements = screen.getAllByRole('button');
